refactor(socket.io): deprecated SocketModules.chats.searchMembers in favour of api.search.roomUsers
