+++ src/socket.io/modules.js
@@ -157,51 +157,16 @@ SocketModules.chats.sortPublicRooms = async function (socket, data) {
 };
 
 SocketModules.chats.searchMembers = async function (socket, data) {
+	sockets.warnDeprecated(socket, 'GET /api/v3/search/chats/:roomId/users?query=');
+
 	if (!data || !data.roomId) {
 		throw new Error('[[error:invalid-data]]');
 	}
-	const [isAdmin, inRoom, isRoomOwner] = await Promise.all([
-		user.isAdministrator(socket.uid),
-		Messaging.isUserInRoom(socket.uid, data.roomId),
-		Messaging.isRoomOwner(socket.uid, data.roomId),
-	]);
 
-	if (!isAdmin && !inRoom) {
-		throw new Error('[[error:no-privileges]]');
-	}
-
-	const results = await user.search({
-		query: data.username,
-		paginate: false,
-		hardCap: -1,
-	});
-
-	const { users } = results;
-	const foundUids = users.map(user => user && user.uid);
-	const isUidInRoom = _.zipObject(
-		foundUids,
-		await Messaging.isUsersInRoom(foundUids, data.roomId)
-	);
-
-	const roomUsers = users.filter(user => isUidInRoom[user.uid]);
-	const isOwners = await Messaging.isRoomOwner(roomUsers.map(u => u.uid), data.roomId);
-
-	roomUsers.forEach((user, index) => {
-		if (user) {
-			user.isOwner = isOwners[index];
-			user.canKick = isRoomOwner && (parseInt(user.uid, 10) !== parseInt(socket.uid, 10));
-		}
-	});
-
-	roomUsers.sort((a, b) => {
-		if (a.isOwner && !b.isOwner) {
-			return -1;
-		} else if (!a.isOwner && b.isOwner) {
-			return 1;
-		}
-		return 0;
-	});
-	return { users: roomUsers };
+	// parameter renamed; backwards compatibility
+	data.query = data.username;
+	delete data.username;
+	return await api.search.roomUsers(socket, data);
 };
 
 SocketModules.chats.toggleOwner = async (socket, data) => {
